Improved QEKF and Velocity estimator + Working velocity control for V1
First time the velocity controller is well working for Kugle V1

+++ KugleFirmware/Libraries/Modules/Estimators/COMEKF/COMEKF.h
@@ -42,7 +42,7 @@ class COMEKF
 		Timer * _microsTimer;
 		uint32_t _prevTimerValue;
 
-		int32_t _prevVelocity[2];
+		float _prevVelocity[2];
 
 		/* State estimate */
 		float X[2];   // state estimates = { COM_X, COM_Y }
